feat: add module documentation and rename
